server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "user_management"}]

[tool.poetry.dependencies]
python = "^3.10"
django = ">=5.2.1,<6.0.0"
django-prometheus = ">=2.3.1,<3.0.0"
prometheus-client = ">=0.21.1,<0.22.0"
celery = ">=5.5.2,<6.0.0"
django-celery-results = ">=2.5,<3.0"
djangorestframework = ">=3.14.0,<4.0.0"
psycopg2-binary = ">=2.9.9,<3.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

